refactor `read_many` to take a stream of iovecs as input
Instead of taking an iterator of (offsets, positions), make it a stream.

The reason behind doing that is that using an iterator forces us to
schedule all the IO synchronously at call-time (or keep them in memory
which isn't great either). While this is fine when dealing with a small
number of requests (and -marginally- more efficient) it becomes
problematic as the number of IO requests increases.

Specifically doing this may starve other tasks performing IO in the same
task queue since the reactor submits IO requests to the devices in FIFO
order. Therefore, if two concurrent tasks call `read_many` they will
race and all the IO submitted by the first task will be processed before
a single IO from the second reaches the IO ring. This is obviously wrong
as we want some amount of fairness even within a common task queue.

Another use case is when performing parallel calls to `read_many` with
the invention of zipping the resulting streams. For instance, if some
data is spread in many files we may want to fetch data in parallel and
combine the result asynchronously.

One caveat of the current state of the code is that the stream will
create only one `Source` at a time. Therefore, this refactoring
temporarily removes all IO concurrency within a `read_many` invocation.
The next commits will reintroduce concurrency.