try to use a skip list to store timers to avoid split allocation
The existing routine that consumes timers allocates a lot because of the
b-tree splitting. Since we try to consume timers more often than we
create them, it makes sense to optimize for the former at the expense of
the latter. A skip list should provide fast insert speed and very cheap
draining.

I chose to use a skip list, so I had to add a new dependency. We can
fall back to a linked list if that's not acceptable.