add a 1-element LIFO buffer in front of the local task queue
This is a trick from the Go scheduler to get most of the benefits of a
LIFO queue but preserve the fairness we get from a FIFO queue. This
approach eliminates the (potentially large) scheduling latency that
otherwise arises from adding scheduled runnable to the end of the
run queue.

The change works as such:
* scheduled up tasks go in a "priority slot" instead of the queue. If
the slot is already taken, the previous task is replaced and the old one
goes at the back of the queue.
* when the scheduler retrieves the next task, it'll take the task from
the priority slot instead of the queue, if any.
* The `LocalExecutor` maintains an internal counter that is incremented
each time we run a task from the priority slot instead of the queue.
When this counter reaches a constant (256 as of this commit), the task
in the priority slot is pushed at the back of the queue, and we pop the
next task as normal.
* Each time we switch task queues, the priority slot is emptied such
that we pick from the FIFO queue the next time we visit it.

Note that this commit breaks one Glommio unit test, and I removed it.
This unit test wasn't testing for correctness but for the order in which
the reactor run tasks within a task queue. Since we don't make any
guarantee as to that, I feel fine removing it.