add all the fields in `SourceType` necessary to process a request
Today a `Source` doesn't contain all the necessary information required
to serve it. For instance, a `Read` source doesn't include the position
nor the size of the read. That information is contained in the
associated `UringOpDescriptor`.

This setup is weird because it makes the lifetime of requests hard to
follow and makes the reactor logic more complex than it ought to be.
Additionally, it'll make it very difficult to abstract the reactor down
the line (`Source` is generic while `UringOpDescriptor` is an
implementation detail -- some sources aren't served from the rings).

This commit makes sure that every `Source` contains all the relevant
information so that we can get rid of `UringOpDescriptor` in a later
commit.

Also remove some unnecessary logic in the process.