don't assume `Source` has an associated file descriptor
Back when `Source` was introduced, it was only used as a way to
interact with FDs. This is no longer the case, and it feels wrong that
many `Sources` have a hard-coded FD of `-1`.

Instead, when the operation relates to a file, add the `RawFd` in the
`SourceType` alongside the other arguments.

For now, `Source` still exposes a `Source::raw()` function that returns
its associated FD (or `-1` if there's none). That will go away
further along in the patch-set.