make the scheduler part of the reactor instead of the rings
There is now a unique scheduler per reactor, and all sources go through
it. This makes the design simpler because we no longer interact with the
rings directly for ad-hoc purposes like preemption, linking, etc.

Additionally, we now register sources in the source map only when they
enter the kernel instead of as soon as they enter the reactor. Also, we
now systematically register sources in the ap rather than conditionally
(there were many exceptions in the code that made it difficult to know
what was and wasn't in the map).

This commit disables rushing to the ring for now. It'll be re-enabled in
a further commit.