add a callback system to track source lifecycle
In a further commit, I will need to keep track of the number of inflight
IO requests to rate-limit them. The best way to do this elegantly is to
add a generic way to add callbacks to sources.

Callbacks aren't cheap, but Sources come with no callbacks by default,
so you don't pay for what you don't use. Later on, we'll want to convert
the stats collection logic to use callbacks instead so that it may live
outside the reactor.